remove duplicated file for translator module: moved in module file
