+++ src/main.rs
@@ -10,7 +10,7 @@ use command::receiver::receive_user_input;
 use command::validated_options::validated_user_inputs::ValidatedCommandOptions;
 use std::process::exit;
 use translator::translation_error::TranslationError;
-use translator::translator::TimezoneTranslator;
+use translator::TimezoneTranslator;
 use validator::command_options_validator::validate_command_options;
 
 fn main() {
